+++ concordium-sdk/src/main/java/com/concordium/sdk/crypto/wallet/Network.java
@@ -1,0 +1,14 @@
+package com.concordium.sdk.crypto.wallet;
+
+public enum Network {
+    MAINNET,
+    TESTNET;
+
+    public String getValue() {
+        switch (this.ordinal()) {
+            case 0: return "Mainnet";
+            case 1: return "Testnet";
+            default: throw new IllegalArgumentException("An unsupported network was provided");
+        }
+    }
+}